feature: create single http server and doctor result
Previously: For each workspace folder a separate http server was created and a separate doctor result was produced.
Now: Only a single http server is created and information from doctors is collected into a single doctor result.
