+++ metals/src/main/scala/scala/meta/internal/metals/PopupChoiceReset.scala
@@ -7,7 +7,6 @@ import scala.meta.internal.bsp.BspConnector
 import scala.meta.internal.bsp.BuildChange
 import scala.meta.internal.metals.MetalsEnrichments._
 import scala.meta.internal.metals.clients.language.MetalsLanguageClient
-import scala.meta.internal.metals.doctor.Doctor
 import scala.meta.io.AbsolutePath
 
 import org.eclipse.lsp4j.MessageActionItem
@@ -18,7 +17,7 @@ class PopupChoiceReset(
     workspace: AbsolutePath,
     tables: Tables,
     languageClient: MetalsLanguageClient,
-    doctor: Doctor,
+    executeRefreshDoctor: () => Unit,
     slowConnect: () => Future[BuildChange],
     bspConnector: BspConnector,
     quickConnect: () => Future[BuildChange],
@@ -45,7 +44,7 @@ class PopupChoiceReset(
     } else {
       Future.successful(())
     }
-    result.foreach(_ => doctor.executeRefreshDoctor())
+    result.foreach(_ => executeRefreshDoctor())
     result
   }
